+++ apps/client/src/app/modules/settings/settings-popup/settings-popup.component.ts
@@ -58,6 +58,8 @@ export class SettingsPopupComponent {
 
   metricsPath = '';
 
+  watchFilesPath = '';
+
   proxyType: '' | 'http' | 'https' | 'socks4' | 'socks5' | 'pac' | 'custom' = '';
 
   proxyValue = '';
@@ -164,6 +166,9 @@ export class SettingsPopupComponent {
     this.ipc.on('metrics:path:value', (event, value) => {
       this.metricsPath = value;
     });
+    this.ipc.on('dat:path:value', (event, value) => {
+      this.watchFilesPath = value;
+    });
     this.ipc.once('proxy-rule:value', (event, value: string) => {
       if (!value) {
         if (this.proxyType !== 'pac') {
@@ -215,12 +220,17 @@ export class SettingsPopupComponent {
     this.ipc.send('proxy-bypass:get');
     this.ipc.send('proxy-pac:get');
     this.ipc.send('metrics:path:get');
+    this.ipc.send('dat:path:get');
     this.ipc.send('rawsock:get');
     this.customTheme = this.settings.customTheme;
   }
 
   changeMetricsPath(): void {
     this.ipc.send('metrics:path:set');
+  }
+
+  changeWatchFilesPath(): void {
+    this.ipc.send('dat:path:set');
   }
 
   setProxy({ rule = '', pac = '' } = {}): void {