+++ apps/client/src/environments/patch-notes.ts
@@ -1,55 +1,12 @@
 export const patchNotes = `### Bug Fixes
 
-* **db:** added automated filter for some wrong reports.
-* **db:** fixed reduction data sources for HW nodes.
-* **db:** fixed required item name shown twice in leve details pages.
-* **db:** fixed some monster positions not linked to a proper map.
-* **desktop:** possible fix for update shotcut setting being ignored.
-* **fishing-reporter:** fixed reporter sometimes stopping its reports entirely.
-* **inventory-optimizer:** fixed an issue with "show hidden" not working as intended.
-* **list:** you can now remove team association properly again.
-* **list-panel:** fixed list panel display sometimes breaking in workshop panels.
-* **quick-search:** input field is now properly focused.
-* **recipe-finder:** collectable filter now works as intended.
+* **alarms:** fixed alarms not loading properly in some cases.
+* **desktop:** character detection should now be more reliable.
+* **leveling-equipment:** fixed ring being shown in offhand slot.
+* **list:** fixed layout button not using proper copy mode.
+* **retainer-ventures:** now sorted by total price.
 
 
 ### Features
 
-* **alarms:** added fishing baits in alarm sidebar and overlay.
-* **alarms:** new button to delete all your alarms at once.
-* **alarms:** you can now add an alarm to multiple groups.
-* **alarms:** you can now set "early by" to negative value, making it "late by".
-* **collectables:** you can now create alarms from the collectables page.
-* **company-workshops:** new tree view button in search results and company workshop crafts in lists.
-* **desktop:** inventory tracking now supports multiple characters.
-* **desktop:** middle clicking db buttons now opens the target url in browser.
-* **desktop:** new checkbox in tray menu to control packet capture.
-* **desktop:** new loading window for better loading UX.
-* **desktop:** new tray icon menu entry to reset positions of all your overlay windows.
-* **desktop:** support for command key instead of ctrl for shortcuts on OSX.
-* **gearset:** you can now copy materia and gear piece names in clipboard on click.
-* **gearsets:** new stats breakpoints display.
-* **inventory:** new setting to reset inventory on app startup.
-* **inventory:** new tooltip to display position inside inventory when possible.
-* **inventory-optimizer:** added a reset inventories button.
-* **inventory-optimizer:** new optimizer for useless hq items.
-* **layout:** npc breakdown now takes currency priorities into account.
-* **spearfishing:** new spearfishing spot page.
-* new leveling equipment feature to easily find leveling gear.
-* new retainer ventures page to find best venture for your retainer.
-* new retainers tracking page with alarms for tasks.
-* **list:** new "inventory cleanup" popup to list items inside your inventory that are not used in list.
-* **lists:** new mass delete menu to delete multiple lists at once.
-* **loading-screen:** new patron: G'lek Tarssza.
-* **pricing:** show last updated date on each pricing row when filled from MB.
-* **search:** new dropdown to select language to search with.
-* **search:** you can now exclude a stat or a bonus in search filters.
-* **search:** you can now filter with "Worn by" targeting "Disciple of ..." categories.
-* **sidebar:** new categories to stop having everything in helpers.
-* **sidebar:** you can now set favorite features in settings to list them in the favorites section.
-
-
-### Performance Improvements
-
-* **gathering:** complete refactor of gathering and alarms system.
-* **inventory:** big performance improvements for inventory search operations.`;
+* **desktop:** you can now change the characters folder to watch.`;